chore: move s3 persistence behind flag
