+++ server/Cargo.toml
@@ -24,6 +24,10 @@ eula = false
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ['cfg(tarpaulin_include)'] }
 
+[features]
+default = ["s3-persistence"]
+s3-persistence = ["aws-sdk-s3", "aws-config"]
+
 [dependencies]
 actix-allow-deny-middleware = "0.1.1"
 actix-cors = "0.7.1"
@@ -35,8 +39,8 @@ actix-web-lab = { version = "0.24.1" }
 ahash = "0.8.11"
 anyhow = "1.0.98"
 async-trait = "0.1.88"
-aws-config = { version = "1.6.1", features = ["behavior-version-latest"] }
-aws-sdk-s3 = { version = "1.82.0", features = ["behavior-version-latest"] }
+aws-config = { version = "1.6.1", optional = true, features = ["behavior-version-latest"] }
+aws-sdk-s3 = { version = "1.82.0", optional = true, features = ["behavior-version-latest"] }
 base64 = "0.22.1"
 chrono = { version = "0.4.40", features = ["serde"] }
 cidr = "0.3.1"
