+++ server/tests/s3_tests.rs
@@ -1,4 +1,4 @@
-#[cfg(test)]
+#[cfg(all(test, feature = "s3-persistence"))]
 mod s3_tests {
 
     use std::collections::HashMap;
@@ -11,7 +11,7 @@ mod s3_tests {
     use testcontainers::{ImageExt, runners::AsyncRunner};
     use testcontainers_modules::localstack::LocalStack;
     use unleash_edge::persistence::EdgePersistence;
-    use unleash_edge::persistence::s3::S3Persister;
+    use unleash_edge::persistence::s3::s3_persister::S3Persister;
     use unleash_edge::types::EdgeToken;
     use unleash_types::client_features::ClientFeature;
     use unleash_types::client_features::ClientFeatures;
